CertificateCheckerTest.php
